TinyKey v1.6.8
This is a new version in which it is v1.6.8
It adds alot of new choices,
Makes it so you cannot run on
a video or image,
Checks internal storage for website list
Allows you to add websites not in the current
1000+ list,
Lots of bug fixes,
Added image options,
Etc. Etc.